pbe/meeting-analyzer/requirements.txt
cohere>=4.0.0
python-dotenv>=1.0.0
python-dateutil>=2.8.0
qdrant-client>=1.6.0
sentence-transformers>=2.2.0
requests>=2.28.0
flask>=2.3.0
werkzeug>=2.3.0